breaking: Set docker as base image
Reasoning behind this is that we sometime need to do ugly "local-exec"
within terraform and we sometimes want to have some tools available
(eg: aws-cli). But if we have to install all the tools we need, this
image is going to get very fat. Solution, we propose to mount the docker
socket in the terraform image, and with docker cli installed with
terraform it allows us to use any tools we want.

Furthermore, if you don't want the security risk, just don't mount the
socket. Yes the image does get a little fatter. But that's the price to
pay.